fixed; population of non-schema property
We only need the schema to deduce the model name. If a user
passes the model name we allow it. If we cannot determine
the model name we return an error. This is a change from
< 3.6 where invalid path names were simply ignored. Pointing
out the error to developers is more helpful behavior.

MapReduce benefits:

This means we can now also populate ad-hoc mapReduce results back
to any model of our choice.